搜索功能,产品详情页,购物车,个人中心,登录页,登出功能全部完成

+++ public/front/js/search.js
@@ -10,22 +10,34 @@ $(function(){
     };
     // 清空记录
     $(".history").on("click",".btn-clear",function(){
-        localStorage.removeItem("searchList");
-        render();
+        // 弹出框
+        mui.confirm("你确定要清空历史记录吗","温馨提示",["取消","确认"],function(e){
+            if(e.index===1){
+                localStorage.removeItem("searchList");
+                render();
+            }
+        });
     });
     // 删除单条记录
     $(".history").on("click",".btn-del",function(){
-        var arr= getHistory();
-        var index=$(this).data("index");
-        arr.splice(index,1);
-        localStorage.setItem("searchList",JSON.stringify(arr));        
-        render();
+        var that=this;
+        // 弹出框
+        mui.confirm("你确定要删除该条记录吗","温馨提示",["取消","确认"],function(e){
+            if(e.index===1){
+                var arr= getHistory();
+                var index=$(that).data("index");
+                arr.splice(index,1);
+                localStorage.setItem("searchList",JSON.stringify(arr));        
+                render();
+            }
+        });
+        
     });
     //添加记录
     $(".search .search-btn").click(function(){
         var key=$(".search .search-input").val().trim();
         if(key===""){
-            alert("请输入搜索关键字");
+            mui.toast("请输入搜索关键字",{ duration:2000});
             return;
         }
         var arr=getHistory();
